OrganizationTestsMenu improvements: Ph.1
### What's done:
* Replaced react-tables with custom lists on test tab of OrganizationView
### TODO:
* Add TestSuiteSource deletion
* Improve buttons on TEST tab
* Add menu for permissions browsing

(#1097)

+++ save-frontend/src/main/kotlin/com/saveourtool/save/frontend/components/basic/organizations/OrganizationTestsMenu.kt
@@ -20,6 +20,8 @@ import csstype.ClassName
 import react.*
 import react.dom.html.ReactHTML.div
 
+import kotlinx.browser.window
+
 /**
  * TESTS tab in OrganizationView
  */
@@ -117,9 +119,11 @@ private fun organizationTestsMenu() = FC<OrganizationTestsMenuProps> { props ->
         testSuitesSourceUpsertWindowOpenness.openWindow()
     }
     val deleteHandler: (TestSuitesSourceSnapshotKey) -> Unit = {
-        setTestSuitesSourceSnapshotKeyToDelete(it)
-        deleteTestSuitesSourcesSnapshotKey()
-        setTestSuitesSourceSnapshotKeys(testSuitesSourceSnapshotKeys.filterNot(it::equals))
+        if (window.confirm("Are you sure you want to delete snapshot ${it.version} of ${it.testSuitesSourceName}?")) {
+            setTestSuitesSourceSnapshotKeyToDelete(it)
+            deleteTestSuitesSourcesSnapshotKey()
+            setTestSuitesSourceSnapshotKeys(testSuitesSourceSnapshotKeys.filterNot(it::equals))
+        }
     }
     val (managePermissionsMode, setManagePermissionsMode) = useState<PermissionManagerMode?>(null)
     manageTestSuitePermissionsComponent {
